refactor: extract local copy context module
